Fix partial layer packages in installed models
Fixes #1004
